Don't have separate deques for queued and written messages
Summary: When reading we need multiple deques because the messages hop between the pipe (when waiting for the descriptor), then the user (when waiting for allocation), then the pipe again (when reading). For writes, the two queues just differed about whether we had sent out the writes yet or not, so it makes sense to merge them and just add a boolean flag.

Reviewed By: jiayisuse

Differential Revision: D20139941

fbshipit-source-id: 86456b31062ee9f1b6dd05e2ffd9e6d0f025753f