Separate pipe into pipe+impl
Summary:
Another preparatory diff to fix the problem described in the previous diff.

This is the first step towards having a separate lifetime for the pipe (the user-visible "handle") and the internal implementation (whose lifetime will be different as it may need to survive for longer than the pipe in order to wait for asynchronous callbacks to complete).

For now, however, we just separate them into two classes, while still keeping their lifetimes tightly coupled.

Reviewed By: jiayisuse

Differential Revision: D20522370

fbshipit-source-id: c8dfd4a323676a31f876b34ea50f7a644db6b7ae