Extend lifetime of pipe when needed by callbacks
Summary:
This diff is a preparatory step for some re-working of the pipe lifetime and its callbacks. The core problem we're trying to solve is the following: we have a commitment to always invoke the callbacks the user passes to us (possibly with an error), including when the pipe gets closed (by destroying its last shared_ptr). However, there may be background threads working on that data (In the transport/channel) and we thus need to wait for them to terminate before we can safely invoke the callbacks. (The risk here is that if we invoke the callbacks while some thread is working on the data, we signal to the user that we're done with the data and they may deallocate it, which could lead to segfaults in the other threads). Since the transports/channel tell us when they are done with the data by invoking a callback, we need to keep the pipe alive until those callbacks are invoked. This is what we do here.

If you remember the rationale behind runIfAlive, this new approach seems to go against it. It is indeed the case, because that solution turned out to not cover all the cases. So instead of having runIfAlive do an opportunistic call based on whether the object is alive, we are going to explicitly control (and extend) the lifetime of the object.

This diff leaves the pipe in a broken state. Namely, after this diff, if one schedules some operation on a pipe (a read, a write, ...) that never completes, one can never really destroy the pipe, as it will be kept alive by that pending callback. A fix for this will come later in the stack, and will consist in detaching the pipe from its implementation, and ensuring that destroying the pipe will trigger a (delayed) shutdown of the implementation.

Reviewed By: jiayisuse

Differential Revision: D20522368

fbshipit-source-id: 17fca4d8d0d4fc8b2a72b65492383ef1a847388d