Don't invoke pipe callbacks when data is still being worked on
Summary: This diff fixes the problem I mentioned in earlier diffs in the stack: we shouldn't call the callbacks while there are still background threads working on the (as then the user may deallocate it and those threads might segfault). So instead we must wait for all transport/channel callbacks to complete before invoking the pipe callbacks. Which is what this diff does.

Reviewed By: jiayisuse

Differential Revision: D20522366

fbshipit-source-id: aeac95b56a4d6de882285490de7a1cf05ff70195