Have channels return descriptor asynchronously
Summary:
This diff is in some sense the start of a new stack, although it's closely related to the previous ones: we've reworked the thread model for the transports and for the pipe, we now need to do it also for the channels. We've soon hit a snag: we cannot defer the send operation if the caller expects it to synchronously return a result (the tensor descriptor). Thus we need to change it so that the descriptor is returned using a callback.

This may also prove to future-proof the channel API since it's not unlikely that some channels, for example the CUDA one, may need to perform some expensive operation in order to prepare the descriptor, and would thus prefer to do so asynchronously.

This moreover makes it also easier for us to return an error if producing a tensor descriptor fails (which is certainly something that can happen for, like, CUDA).

There shouldn't be a performance hit because even though we say the callback is "deferred" it is in practice just appended to the task queue of the current temporary event loop, thus it will be run immediately after, without context switches. After these changes, though, the exact same operations will be performed by the exact same threads, just in a possibly different order.

Reviewed By: jiayisuse

Differential Revision: D20559969

fbshipit-source-id: 365b96042734b9751c6e171018fb83415c16ba16