Fix lock inversion bug
Summary:
There was a bug in the refactoring of the threading model, which only surfaced once both the pipe and the channel were ported to the new design. The bug boiled down to the fact that the temporary event loop thread kept holding the lock while executing the pending tasks. This could lead to a lock inversion deadlock, when there were two threads, both of which needed to run operations/callbacks on both the pipe and the channel, but did so in different orders. Here is an example:
- the user thread could call write on a pipe, which would call send on a channel
- the read callback of a channel may fire, which in turn would trigger the recv callback on a pipe.

The solution is simple: instead of using a held lock to signal that there's a temporary event loop already running, just set a flag (i.e., set `currentThread_`), and only acquire the lock when one needs to change that flag or to change the pending tasks.

This leads to a small functional difference: before, if a thread wanted to run an operation/callback on an object (pipe or channel) and that object already had an event loop running, the thread would wait for that event loop to finish, and then run the task itself. Now, instead, the thread will just offload its task to the event loop that is already running. I think this may actually help performance because it will avoid one thread going to sleep and then being woken up.

Reviewed By: jiayisuse

Differential Revision: D20619766

fbshipit-source-id: a0830c38f077395424a328ade3ecf061534b3d58