Explain the new thread model
Summary: This is a brain dump of the process that led us to our current design around threads. It ended up being a bit long but I wanted to include enough context (examples of issues we had to solve, failed attempts, ...). This should help as reference, to help onboarding and as a checklist for any proposed simplification.

Differential Revision: D20669291

fbshipit-source-id: 1ab74c56435b1df300ae4eea48e2016e0c75cd12