Explain the new thread model (#138)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/138

This is a brain dump of the process that led us to our current design around threads. It ended up being a bit long but I wanted to include enough context (examples of issues we had to solve, failed attempts, ...). This should help as reference, to help onboarding and as a checklist for any proposed simplification.

Reviewed By: osalpekar

Differential Revision: D20669291

fbshipit-source-id: 16746f14e8fc6ea52fc0310c11587de285c7b646