Fix tensor-less messages "overtaking" tensor-ful ones when writing
Summary:
In D20559969 we made channels provide tensor descriptors by calling a callback. This means that messages with tensors now require at least two loop iterations to be sent. Messages without tensors though are sent right away. This could mess up the ordering. The fix is to keep them in the queue until all the ones before have also been sent.

Thanks to jiayisuse for finding this and providing a test case.

Reviewed By: jiayisuse

Differential Revision: D20769031

fbshipit-source-id: dea1b12dc8276fc8d7c9acb4403e2fc4ea9d522a