Add close methods to transports
Summary:
This diff adds clearly defined methods to the transport objects to perform graceful shutdown. All objects can be closed (which happens implicitly when destructing them) and the contexts can also be joined (also happens when destructing).

This diff mostly prepares the API and connects it to whatever code was already there for that purpose, even if in some cases that code may not reflect the exact semantics we want. (E.g., closing a context doesn't currently always close all the connections). This will come later.

Reviewed By: jiayisuse

Differential Revision: D20677489

fbshipit-source-id: 924cea619fcd456bcd21539b2a13d5b7d58cf441