Fix Python tests not releasing the GIL when joining
Summary:
With the following stack, we'll finally get the closing/joining/destroying behavior right. This causes a problem in our simple Python bindings because of the GIL: pybind11 is holding the GIL while destroying the context, but this may fire some callbacks, that need the GIL. There's no easy way to release the GIL during destruction, so for now the solution is explicitly calling join (and make sure we release the GIL there).

Also, the actual bug was caused by the fact that we weren't waiting for all the write callbacks to complete. This bug got exposed by the introduction of the CMA channel which (contrary to the basic one) fires the recv callback (and thus completes the read) before the send callback (and thus the write).

Reviewed By: beauby

Differential Revision: D20723638

fbshipit-source-id: e1ce74f9bba8270ab272827689b59353dd0bac53