Ensure closing the context/factory closes all pipes/connections/listeners/channels
Summary: We recently decided that closing a context or a factory should in turn cause all of its dependent objects to become closed too, namely pipes, connections, listeners and channels. So far this was only happening in the UV transport using some ad-hoc system. This diff introduces two generic classes in callback.h that make it very easy to add this behavior everywhere, which is what we do. (With the exception of the SHM and the TLS transports, as they still have to be ported to the new design).

Reviewed By: beauby

Differential Revision: D20710534

fbshipit-source-id: 1bee04173ea3a59ed81af916310cfbbe4e3ffcba