Split listener into listener+privateiface+impl
Summary:
Splitting the public object from the private impl is something we've done a thousand times before. But this time there's something more.

The pipe needs to have access to the listener in order to establish additional connections, if it needs to switch to a more performant transport or if it needs to open channels.

Until now, the pipe was holding a shared_ptr to the listener and was accessing its private methods (as the pipe was marked as a friend of the listener). This had several problems:
- By keeping a shared_ptr to the listener the pipes were violating the "single real owner" principle and causing the listener to remain alive even after the user stopped using it.
- By directly accessing its private members and calling its private methods, the pipe's usage of the listener could eschew the access patterns required by the listener (locks, event loops, ...).

By defining an abstract interface for the pipe to use, by having the impl adhere to it, and by having the pipe store a shared_ptr to the impl while thinking it's a shared_ptr to that interface we solve the above problems.

Reviewed By: patricklabatut

Differential Revision: D20710532

fbshipit-source-id: 276cc186c5a51fdb9b98d08b2b05c92f260c42d7