Move core impls to .cc files
Summary:
This is the cherry on top of all the effort we did around splitting objects into a public shell, a private interface and a private implementation. As the implementation definition is now private, it can be moved from the .h to the .cc file, thus enforcing no external access.

This diff introduces no functional changes.

Reviewed By: patricklabatut

Differential Revision: D20868058

fbshipit-source-id: a60ead3c51cb0ea4ee9ea88aa22b3c33ce807c60