Split CMA factory into factory+privateiface+impl
Summary:
Each channel is currently keeping a shared_ptr to the factory, which means that when the user destroys its pointer to the factory the destructor may not get called and the factory may remain open and active. This goes against the semantic we've adopted everywhere else (i.e., if something goes out of scope it will be cleaned up). In order to achieve the right behavior we need to split the factory up into a public object (held only by the user) and an impl (which is what channels have a pointer to): this way we can detect the public object being destroyed and can in turn shut down the impl.

As a separate problem, the channels are currently accessing the private internals of the factory without proper control. Introducing an impl would help, but the impl remains a private implementation detail, which we'd like to move to the .cc file. To properly address that we introduce a privateiface abstract class, that the impl will subclass, which tells the channels what operations they can perform on the impl.

Reviewed By: beauby

Differential Revision: D20773340

fbshipit-source-id: 3fc54a76a8f8761da2be4e6febdba73d1cd93953