Consolidate UV connection/listener initialization
Summary:
Creating a listener or a connection on a given address involved many steps:
- The string was converted to a Sockaddr by the context's methods.
- A libuv handle was created either in the factory object of the public object or in the constructor of the impl.
- The initialization of that handle was performed using ad-hoc deferToLoop calls.

This diff streamlines all of the above by having everything be done inside the impl constructor or the initFromLoop method (if it has to be run from the loop). All other methods will just forward the arguments without doing anything with it. In passing I also removed the factory methods as they don't serve much of a purpose anymore.

Reviewed By: beauby

Differential Revision: D20898215

fbshipit-source-id: ae29ab19832f17e805bdcb4bd8927cb090c53397