Consolidate SHM connection/listener initialization
Summary:
Creating a listener or a connection on a given address involved many steps:
- The string was converted to a Sockaddr by the context's methods.
- A socket was created either in the context's methods or in the constructor of the impl.
- The initialization of that socket was performed either in the constructor (thus outside of the loop!) or in initFromLoop.

This diff streamlines all of the above by having everything be done inside the impl constructor or the initFromLoop method (if it has to be run from the loop). All other methods will just forward the arguments without doing anything with it. In passing I also removed the factory methods as they don't serve much of a purpose anymore.

Reviewed By: beauby

Differential Revision: D20898213

fbshipit-source-id: e4a1316ea59523d2dc8c78e95f37a2a57abde677