Remove superflouous enable_shared_from_this
Summary:
There were a bunch of classes that inherited from enable_shared_from_this without needing it. Some of them were the core context and listener, which now are wrappers around an impl, and only the impl needs to retrieve its own shared_ptr. Two other ones were the loops of SHM and UV, where shared_from_this was used but could be avoided. In particular, it turned out that the SHM loop's monitor was unused, so I removed it. But, actually, since it came with a test and it's a shame to remove test code, I just moved the monitor to a helper function in the test.

Now only impls extend from enable_shared_from_this.

Reviewed By: beauby

Differential Revision: D20941788

fbshipit-source-id: 3174e4d6848474de6c8df407d040ce6edd7bb4ba