Simplify object construction across the board
Summary:
We used factory methods to make sure we always called init after constructing object, as relying on the users to always do so isn't robust. We used constructor tokens (a.k.a., the passkey idiom) to hide these constructors and only make them available to the factory methods (or to the friend classes).

Having now moved everywhere to impls, some of these issues have naturally disappeared: as the impl class is defined in .cc files, only code in that .cc file can construct it. By having this very strong enforcement on the users of the impls, we can relax our other containment measures.

Another benefit of impls is that even if they need a separate init method for setup, this doesn't extend to their public wrapper. What I mean is that there is no need for an init method on the wrapper, which can instead be constructed with just the constructor, and it will be the constructor to call the impl's init method. This allows us to remove some factory methods for the public wrappers, and the associated constructor tokens.

Reviewed By: beauby

Differential Revision: D20898214

fbshipit-source-id: 703e3d4c1baa60c4a4d3db124799828a4ecda093