Support 0-sized reads and writes in UV
Summary: The over-the-wire format of zero-sized chunks is 8 bytes containing the size (hence all zeros) and nothing more. Since this is the same format as for the rest, very few changes are necessary. The only reason we need this is because libuv was returning ENOBUFS errors when we passed empty buffers to it.

Reviewed By: jiayisuse

Differential Revision: D21023073

fbshipit-source-id: 1d7a9f36cbe64a72e9341f577c5ca93fed546c40