Fix expired reference in listener
Summary: When calling all the listener's callbacks with an error, we use a lambda to generate the arguments of each invocation. This was creating a temporary tuple that held a copy of the value of the error, which was however then converted to a reference (as that is the type expected by the callbacks). This was a reference to the temporary, which got destroyed.  Instead, we should just return a reference to the original error, which will be kept alive.

Reviewed By: jiayisuse

Differential Revision: D21023074

fbshipit-source-id: eb3d683d54377e3c643c87ecdfc9f88d27646e7a