Reduce reliance on shared_ptr for SHM loop and reactor
Summary:
This diff's title isn't great, but here is what it means: we used to force the SHM loop to be a shared_ptr, by having a factory method (which didn't do anything more than call make_shared).

The loop also stored a shared_ptr to the reactor, even though there were no other users to share that pointer with. A regular member value would have worked just fine.

This diff removes the factory method and exposes the constructor, so that users can create a loop in the way they want (this will be useful in the next diff). It also makes the loop store the reactor by value, rather than via a shared_ptr.

Reviewed By: beauby

Differential Revision: D20963680

fbshipit-source-id: 13e3d836042e9ff18e93a2e3ceca6fbb6d5fb30f