Have the SMH impls do the deferring to loop, instead of the public objs
Summary: The goal of this diff is to avoid having the logic split between the public object and the impl, where the latter relies on the former to defer all the calls to the loop. After this diff, the public object will be a dumb wrapper that does nothing but forward, while the impl will be a self-contained object. The result is nice encapsulation.

Reviewed By: beauby

Differential Revision: D20961694

fbshipit-source-id: ccbf01b4048dc98f894d9c669a6b96df00ad9ddb