Reduce reliance on shared_ptr for UV loop
Summary:
This diff's title isn't great, but here is what it means: we used to force the UV loop to be a shared_ptr, by having a factory method (which didn't do anything more than call make_shared). Moreover, we stored a shared_ptr to it in each handle in order for the handle to keep the loop alive, which is something that was already dealt with by libuv.

This diff removed the factory method and exposes the constructor, so that users can create a loop in the way they want (this will be useful in the next diff). It also makes handles get a reference to the loop rather than a shared_ptr.

Reviewed By: beauby

Differential Revision: D20963682

fbshipit-source-id: 59f1e3e182c4e3b374b7c1adb680f6978a298e61