Introduce a context impl in UV
Summary:
The transport contexts (for both UV and SHM) were the only objects that didn't have an impl class. Except, that's not exactly true: their loops behaved in pretty much the same way a context impl would. The difference was therefore just a matter of naming.

This diff addresses such inconsistency by introducing a real context impl and making the loop a member field of it (by value, not by shared_ptr). This allows us to clearly expose in the context's header file what interface the connection and listener need. It also completely hides the loop and makes it an implementation detail.

Reviewed By: beauby

Differential Revision: D20963681

fbshipit-source-id: be62dbec3687a5d11f9b1e78b595ef9fe944deff