Give names to threads
Summary: For easier debugging of production runs we can give names to threads that will then show up on GDB (and others), making it easier to find out threads and get stack traces. The standard library doesn't expose this feature, so we must use pthread directly, and one of its non-portable (`_np`) calls nonetheless. However, since this is just for debugging, it's fine to gate it to only work on Linux and do nothing in other cases.

Reviewed By: jiayisuse

Differential Revision: D21130254

fbshipit-source-id: 75823625c3d33c10347a7974f9b435bd1981d629