Allow users to provide semantically meaningful names for better logging
Summary: In the previous diff we gave unique auto-generated names to contexts, listeners and pipes to identify them in logs. These names may give little insight to users, who are more interested in the addresses or even the roles of processes in the system. This diff enables the latter, by allowing users to give each context a name and by specifying the expected name of the target process when opening a pipe to it. This allows us to know the names of the source and target processes of each pipe (both the outgoing ones and the incoming ones). We print those pairs for each pipe ID, allowing an easier interpretation of the logs.

Reviewed By: jiayisuse

Differential Revision: D21130253

fbshipit-source-id: 7d70c20323c40c636851767b7a663a3a2e75594d