Add tons of logging to trace flow in pipe
Summary:
Sometimes some tricky issues seems to only come up in production runs and are hard to get a small repro for. Being able to figure out what's going on by working with the full production system is important. So here I am adding logging to be able to follow the logic of the pipe and the other high-level components. I am basically adding calls:
- when operations are performed on those objects by the user
- before *and* after control is deferred to someone else and we're waiting for them to give it back (either when the pipe starts an operation on, say, the connection; or when the pipe calls a user callback)
- when objects are destroyed

As this is very verbose and we don't want it to be always on, I'm gating it behind a `TP_VERBOSE_LOGGING` preprocessor directive. I'm not `TP_DLOG` because we may want to enable those logging also in non-debug builds.

Reviewed By: jiayisuse

Differential Revision: D21130252

fbshipit-source-id: cd758a0b3e2bdb9b38afd191a04bc8d5ae55c9e6