Use string for tensor descriptor instead of vector<uint8>
Summary: The two representations are essentially equivalent, especially since we don't manipulate the data but just store it and move it. Converting between the two however is rather inefficient, because we cannot move and have to copy. I see no reason not to use strings and make our life easier.

Reviewed By: jiayisuse

Differential Revision: D21200472

fbshipit-source-id: a3c0f3a4af80cdef3dd4e3870f5ae56c3a356546