Make callback wrappers universal
Summary:
Since all the callbacks we pass around follow the same pattern (store a weak_ptr, lock it back to a shared_ptr, handle errors, ...) we abstracted that away in a CallbackWrapper class. We then created an instance of this class for each type of callback. Well, it turns out we don't need that, and we can do it with just one instance.

To achieve that I'm making heavy use of templates, universal references and forwarding, auto return type, decltype, and so on. Am I proud of it? Yes. Does it feel dirty? Also yes.

In future diffs I'm considering:
- Avoiding the callback wrappers being friends of the other classes, and instead have these classes pass to the wrapper's constructor references to all the members it needs.
- Making all callbacks eager: if the parent object gets destroyed it first gets closed, and then it closes all the child objects, which thus fire the callback. It therefore shouldn't change the behavior but it should simplify the logic quite a bit and make it easier to reason about the code.

Reviewed By: jiayisuse

Differential Revision: D21131647

fbshipit-source-id: bdb4e227ccac184eb34de8639ea2782e2d9fe028