Allow messages to contain vector of payloads, instead of single one
Summary:
By only allowing a single payload we're requiring users that want to transfer multiple pieces of data at the same time (because they are logically related?) to copy them into a contiguous buffer. TensorPipe is against useless copies, so we should allow multiple payloads to address that.

This feature is a breaking change in the API and in order to avoid racing with the PyTorch RPC agent I will first merge this diff which just adds the option of multiple payloads while keeping the old code. I will then update the agent (either before or after it is merged). Once the agent is updated, I will remove the old API.

Reviewed By: jiayisuse

Differential Revision: D21209901

fbshipit-source-id: 1fb48fe48a986ee6daa728f1847e970cfac187ed