Fix anew PyTorch build integration
Summary:
Even after I thought I had fixed this (see D21357992) I was still getting errors in PyTorch's CI. In particular `clang-tidy` was failing with this error (see [here](https://github.com/pytorch/pytorch/pull/37729/checks?check_run_id=640154081)):
```
CMake Error: install(EXPORT "tensorpipe-targets" ...) includes target "tensorpipe" which requires target "uv" that is not in any export set.
```
(I hadn't noticed that before before the agent code was also causing failures in `clang-tidy`)
Note that `clang-tidy` was complaining even though all other jobs were successfully building and running the code. I'm not sure whether it was a false positive...

According to the internet (a Stackoverflow question I cannot find again), the inability to install a target defined in another directory was a limitation of CMake that has been removed only in recent versions, and there was no way around it.

As I was checking how other projects that depended on libuv did to deal with this I stumbled upon uvw (from which our libuv wrappers are inspired) and saw that they were building libuv as position-independent code. This was in fact the original problem I was trying to address by picking the shared version of libuv, so I tried uvw's solution and it seems to work: it makes PyTorch compile and it makes its `clang-tidy` happy.

Reviewed By: beauby

Differential Revision: D21377774

fbshipit-source-id: aa7ca412438e4fc57291b214794318347d36a8bd