Use classes to provide options to constructors
Summary: Before we start using this API in PyTorch, let's make it a bit more future-proof by putting all the optional arguments in an "options" class. This is called the builder pattern in C++, it is the same one used by PyTorch for TensorOptions, and it sort-of allows to recreate the keyword arguments from Python.

Reviewed By: jiayisuse

Differential Revision: D21479803

fbshipit-source-id: a4374d91f58abdfd194da0e777436f30bc3a9ab6