Clean up UV error handling
Summary:
The listener didn't have error support at all. (Meaning when it received an error from libuv it failed the first callback, but not the subsequent ones).

The connection had some cleanup code all over the place which is now neatly gathered in the error handling function. Also, in some cases it was overwriting the error.

Reviewed By: beauby

Differential Revision: D21325946

fbshipit-source-id: bb841ae16c33f65b0094eed350ff435da1e246c5