Auto-generate ids for transport listeners and connections
Summary: In the last diff we gave all contexts an id, it's time now for those contexts to propagate this id to the objects they create (connections, listeners, channels). Our naming scheme is hierarchical, with the core context being the root and each new level appending a component to the name. There is a "diamond scheme" here, i.e., there are two paths going from the core context to, say, a connection: one going through the transport's context, the other going through the pipe. They give different names. However, we don't always have the name that comes from the pipe's name, because sometimes connections don't belong to pipes. So I think the default name of connections should be the one coming from the transport's context, and then the pipe can override this name. The code to do this override will come in the next diff.

Reviewed By: jiayisuse

Differential Revision: D21408920

fbshipit-source-id: a09f2df298efcdc6648c126ace0efe06ca0cf649