Don't have channels be aware of their name
Summary: We had channels be aware of the name of their class (e.g., `basic` or `cma`) which they were only using for easier debuggability. In the next diffs I'll add ids to channels that will contain both their type and more information like which pipe they belong to, so those ids should be preferred to the names. Therefore I'm removing the names here to avoid duplication.

Reviewed By: jiayisuse

Differential Revision: D21408923

fbshipit-source-id: 7d291ec48725e71e1861e1fcfc1a116b8018050f