Encapsulate the on-demand loop pattern in its own class
Summary:
Several objects were using the on-demand loop design, which lead to the same code repeated over and over. By putting all that code in its own class we can avoid the duplication.

As an added bonus, by giving a reference to this loop class to the callback wrappers we remove the last requirement of having these callback wrappers be friends of the pipe, the listener, the channels, ...

Reviewed By: beauby

Differential Revision: D21325948

fbshipit-source-id: e88b7515019886c00dba398631fac4f46fc1ea25