Fix deadlock in closing emitter
Summary:
The closing receiver creates a lambda that captures a weak_ptr to some object and locks it to a shared_ptr when invoked (using runIfAlive). From debugging a deadlock it appeared that in case of multiple threads operating on an object, that temporary shared_ptr may sometimes be the last one for that object, thus causing the object to be destroyed when the shared_ptr goes out of scope. The destructor of the object calls the destructor of the closing receiver which in turn calls the `unsubscribe` method on the closing emitter. However, when this happens, we're already inside the `close` method of the emitter. Both those methods want to acquire a mutex, and thus they deadlock.

There was already a comment hinting at the fact that such a mutex wasn't an ideal solution, and that we should instead use an event loop. Now that the on-demand loop is an easy-to-use class (thanks to D21325948), it's trivial to do this fix. (The solution can still be improved, but let's do that another time).

Reviewed By: beauby

Differential Revision: D21499446

fbshipit-source-id: f1f8b4839d11332351f56a6829becdd5566111b9