Have closing use the same code path as errors
Summary:
The way all of our objects should handle errors is the following:
- They all start in a "success" state
- As soon as they encounter an error (in their own operations or from a callback invoked from somewhere else), their transition to an error state (which is terminal: they can never recover from it and they are unable to do anything while in it). The error of the object (type, message, ...) is set to that error.
- They clean up once they get into that error state, releasing resources, aborting operations, ...
- Any further error will have no effect, it will not overwrite the previous error (type, message, ...).

Closing an object turns out to be largely the same: it also puts it in a terminal state from which it can't recover and in which it can't do anything. Thus we implement closing as setting to a specific error (i.e., FooClosedError).

Until now the above behavior were inconsistently applied: sometimes we wouldn't clean up after error, or we'd have a separate cleanup after closing, or we'd overwrite previous errors with later ones.

This diff introduces a ErrorHolder class (better name suggestions welcome!) which wraps an error and limits the ways in which it can be accessed. In particular it prevents it from being overridden and once it's first set it calls a cleanup function provided at construction time.

The diff then proceeds to use this ErrorHolder everywhere (except transports, which are trickier and will come later).

Reviewed By: beauby

Differential Revision: D21325949

fbshipit-source-id: 2ebcec58595fa07e7cbef72e0864c3def83ccbf4