Fix race condition in UV loop closing
Summary:
The closing of the event loop thread has had a known race condition for a while, which I'm fixing here. This was causing some deferred functions to never be run, which could lead to memory leaks (because of reference cycles) or to crashes (because of failed asserts).

The problem was simple: some objects were trying to defer functions to the loop once this has been closed and joined. This could reasonably happen even under intended usage. Imagine for example that the user had a context and a connection, both alive and running. It then explicitly joined the context, which would shut down the event loop, closing the connection in passing. The user would then exit the scope, causing both the connection to be destroyed. The connection's destructor cannot know that the connection has already been closed, because such information is stored on the impl, and we can only access data on the impl from within the loop (this restriction is what's keeping us save from race conditions). Therefore the only thing the connection can do is defer to the loop a task that closes the impl. Since the loop is closed, that task will stay there forever, and will hold a reference to the impl, keeping it alive.

We already had some measure in place to have the loop consume as many pending tasks as possible before terminating, but this wasn't enough. (In particular, it wasn't covering the situation above). The only robust solution I could find is detecting if the loop is dead when deferring and, in that case, do so inline. In fact, not *really* inline as that might lead to the sort of reentrant locking that has caused us so many problems. We should use the on-demand loop, which is the solution we have found to the reentrant calls.

Reviewed By: beauby

Differential Revision: D21494695

fbshipit-source-id: 699b15bd723883e1dddb7f1d2e5d6df579385bdc