Fix race condition in channel tests
Summary: If, say, the send operation finished before the recv was complete, the sending thread would finish and destroy its end of the channel, possibly causing an error on the recv side. The inverse could also have happened. So we should have each side wait for the other to finish before closing its side of the channel.

Reviewed By: jiayisuse

Differential Revision: D21409355

fbshipit-source-id: d9813b31c6827b9f5e25f7e2f015c49aba38e8bb