Have UV's Sockaddr take a pointer instead of a reference
Summary: The Sockaddr constructor takes a reference to a `sockaddr_storage` and copies its content into its own `sockaddr_storage`. In the next diff we'll be using functions that give us a pointer to a `sockaddr`. The memory backing that pointer may be smaller than the size of a `sockaddr_storage` so, if we cast it to a reference and pass it to the Sockaddr constructor, it may copy unpaged memory, and either overflow or segfault. Instead, the Sockaddr could simply take a pointer to a `sockaddr` and copy only what it needs.

Reviewed By: jiayisuse

Differential Revision: D21495649

fbshipit-source-id: be533324786eaee0d69e699add7ab6c141970338