Add libuv-based helpers to get IP addr of hostname or of given interface
Summary: The default behavior used by Gloo to choose an IP address to bind to is to resolve the machine's hostname. As that sometimes fails, the user can specify an interface name through an env var and then the address of that interface is used instead. For consistency we should replicate that behavior. Thus we're writing helper functions that perform those two operations using only libuv functions in order to be portable.

Reviewed By: jiayisuse

Differential Revision: D21494694

fbshipit-source-id: 120fb71cc2f2dab3a62eebb86e88b0adf980e937