Make verbose logging controlled by a runtime envvar
Summary: Until now, in order to see the `TP_VLOG() << "Foo"` messages one needed to recompile TensorPipe with the `TP_VERBOSE_LOGGING` flag set. This is inconvenient for OSS users of PyTorch as they typically use the pre-built version of PyTorch (and rebuilding PyTorch is a pain...). So it is important to make the verbosity of logging runtime-controlled, for example through an environment variable. This is what we do here.

Reviewed By: patricklabatut

Differential Revision: D21595351

fbshipit-source-id: fbc3c27c9c08a875f4f11236483a7e08b9b626f3