Gracefully handle all socket-related errors in SHM
Summary: Some operations on sockets performed by SHM could legitimately fail, like connecting to a UNIX domain socket address after that socket was closed. Instead of a hard assert, with exceptions and termination and SIGABORT, we should just set the connection to an error state. We do so for all socket operations, just in case.

Reviewed By: beauby

Differential Revision: D21715305

fbshipit-source-id: c5bc09febf5e326e877c369bd4a29544bab5c922