Avoid hard crash in SHM when writing to full reactor
Summary:
In some tests we were hitting an assert triggered by the write to a reactor ringbuffer returning ENOSPC. The reactor ringbuffers have space for 2k tokens but this limit can in fact easily be reached in practice, especially since we've moved much more heavily to defer operations to the loop, each of which writes a token. So this diff addresses the problem in three ways:
- increase the ringbuffer to hold 1M tokens
- do not post a token to the ringbuffer to defer functions, but use a separate mechanism
- retry the write when we get ENOSPC

Each of these approaches mitigates the problem or makes it less likely to occur. None of them fully fixes it however and the latter solution risks creating deadlocks (it's unclear whether a deadlock is a better or worse behavior than a crash). Such a deadlock could occur among multiple processes but that is veeery hard because it requires both of them to fill their ringbuffers at the same time, while they're each trying to write to the other one's ringbuffer. Creating a deadlock within the same process is more probable but much less likely now that we don't use the ringbuffer to defer: it would only happen when creating a connection both of whose endpoints are in the same process. This risk will further be reduced when we introduce the intra-process transport, as in that case the pipe wouldn't even use SHM if both endpoints are in the same process.

Reviewed By: beauby

Differential Revision: D21762966

fbshipit-source-id: 4e7a975feabda02ae525e6139f51702b4a8f95bd