Leverage ordering of operations in basic channel
Summary: The channel interface used to (implicitly) require that channels be able to handle `recv` methods on one endpoint being called in a different order than the `send` methods on the other endpoint. This made them very flexible, but for little gain, since the pipe was always calling them in the same order. However, to provide such a flexibility, they incurred in a heavy performance cost: the "basic" channel, for example, was not so basic because it had the sending end "cache" the data receiving end "poll" for the data, since it had to ensure that the data travelled on the wire in the order the receiving end was expecting it. This means that a transfer required three hops: sending the descriptor, sending the request, sending the reply+payload. If we assume that the recv requests are in the same order as the send, then when sending we can immediately "push" the payload, which results in the whole transfer taking only one hop (as payload and descriptor can be sent simultaneously). This is a great latency improvement. In addition, the code of the channel also gets simplified by a lot.

Reviewed By: beauby

Differential Revision: D21764396

fbshipit-source-id: c8216a67abc556be89179ee55693268d9168c75a